requirements.txt
asyncio~=3.4.3
discord.py[voice]~=2.3.1
spotipy~=2.23.0
wavelink~=2.6.4
yt-dlp~=2023.7.6
